feat: support filtered folder collections
netlify-cms-reader v1.0.4 returns filtered folder collections
without performing the filtering on the actual collections.

For filtered folder collections in Netlify CMS, see:
https://www.netlifycms.org/docs/collection-types/#filtered-folder-collections

This commit implements the support of filtered folder collections

If a folder collection is filtered, then that collection is returned
by applying the criteria configured in its `filter` property.

As the supported types for the filter value is *not* documented,
this implementation accepts all possible types, so a filter
with the setting `{field: tags, value: ['latest', 'most popular']}
is handled, despite the fact that the UI of Netlify CMS (at least
with version 2.10.48) does *not* handle it while configuration
for it is accepted.