feat: Support sortableFields in folder collections
netlify-cms-reader v1.0.4 returns folder collections with
sortableFields set without sorting the actual collections.

For the sortableFields setting in Netlify CMS, see:
https://www.netlifycms.org/docs/configuration-options/#sortablefields

This commit implements the support of the sortableFields setting for
folder collections.

If a folder collection has a sortableFields setting, then that
collection is returned by sorting it in increasing order based on
the first element in the sortableFields array. So if a collection
have a sortableFields setting like `['latest', 'most popular']` then
the collection will be returned with the items sorted by the values of
their `latest` field in increasing order.